Add a simulator executor types package
