+++ simulation/simtypes/README.md
@@ -1,0 +1,7 @@
+# Simtypes
+
+This package defines:
+* Interfaces for Actions
+* Interfaces for Property Checks
+* Interfaces for AppModules to integrate w/ the simulator
+* Interfaces for an App to integrate with the simulator
