+++ simulation/simtypes/action.go
@@ -2,12 +2,9 @@ package simtypes
 
 import (
 	"fmt"
-	"math/rand"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/simulation"
-
-	legacysimexec "github.com/cosmos/cosmos-sdk/x/simulation"
 )
 
 // Action represents a simulator action.
@@ -28,38 +25,9 @@ type Action interface {
 	WithFrequency(w Frequency) Action
 }
 
-type selectActionFn func(r *rand.Rand) ActionsWithMetadata
-
-type weightedOperationAction struct {
-	moduleName string
-	frequency  Frequency
-	op         simulation.WeightedOperation
-}
-
-func (a weightedOperationAction) WithFrequency(w Frequency) Action { a.frequency = w; return a }
-func (a weightedOperationAction) Name() string                     { return fmt.Sprintf("%s: weighted_op", a.moduleName) }
-func (a weightedOperationAction) Frequency() Frequency {
-	return Frequency(mapFrequencyFromInt(a.op.Weight()))
-}
-
-func (a weightedOperationAction) Execute(sim *SimCtx, ctx sdk.Context) (
-	simulation.OperationMsg, []simulation.FutureOperation, []byte, error,
-) {
-	op, futureOp, err := a.op.Op()(sim.GetRand(), sim.BaseApp(), ctx, sim.Accounts, sim.ChainID())
-	// pass in an empty byte array for response data for all weighted operations
-	return op, futureOp, []byte{}, err
-}
-
-func ActionsFromWeightedOperations(ops legacysimexec.WeightedOperations) []Action {
-	return actionsFromWeightedOperations("no module name", ops)
-}
-
-func actionsFromWeightedOperations(moduleName string, ops legacysimexec.WeightedOperations) []Action {
-	actions := make([]Action, 0, len(ops))
-	for _, op := range ops {
-		actions = append(actions, weightedOperationAction{moduleName: moduleName, op: op})
-	}
-	return actions
+type ActionsWithMetadata struct {
+	Action
+	ModuleName string
 }
 
 var _ Action = msgBasedAction{}
@@ -114,31 +82,4 @@ func (m msgBasedAction) Execute(sim *SimCtx, ctx sdk.Context) (
 		return simulation.NoOpMsg(m.name, m.name, fmt.Sprintf("unable to build tx due to: %v", err)), nil, nil, err
 	}
 	return sim.deliverTx(tx, msg, m.name)
-}
-
-func totalFrequency(actions []ActionsWithMetadata) int {
-	totalFrequency := 0
-	for _, action := range actions {
-		totalFrequency += mapFrequencyFromString(action.Frequency())
-	}
-
-	return totalFrequency
-}
-
-func GetSelectActionFn(actions []ActionsWithMetadata) selectActionFn {
-	totalOpFrequency := totalFrequency(actions)
-
-	return func(r *rand.Rand) ActionsWithMetadata {
-		x := r.Intn(totalOpFrequency)
-		// TODO: Change to an accum list approach
-		for i := 0; i < len(actions); i++ {
-			if x <= mapFrequencyFromString(actions[i].Frequency()) {
-				return actions[i]
-			}
-
-			x -= mapFrequencyFromString(actions[i].Frequency())
-		}
-		// shouldn't happen
-		return actions[0]
-	}
 }
